Reduced reproduction for #8 - Dropdown not showing initial value when list is computed
Cause: When the options list is a computed, the value will get assigned before it. This loses the value, because the native select's `.value` setter will change it to '' when it can't find the option (which is always the case here, because they haven't been assigned yet).

The original <select><template is=dom-repeat items=getAirportsForCity(city)></select> doesn't have the issue because templates are evaluated from leaves to roots.